+++ src/main/java/com/rivalia/rivalia/application/port/out/WebClientRepositoryPort.java
@@ -3,13 +3,10 @@ package com.rivalia.rivalia.application.port.out;
 import com.rivalia.rivalia.domain.model.User;
 import com.rivalia.rivalia.infraestructure.outbund.webclient.dto.AuthApiGeneralResponse;
 import com.rivalia.rivalia.infraestructure.outbund.webclient.dto.AuthApiUserSave;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpMethod;
 import reactor.core.publisher.Mono;
 
-import java.util.Map;
-
 public interface WebClientRepositoryPort {
-    Mono<AuthApiGeneralResponse<AuthApiUserSave>> saveInAuthApi(User user);
+    Mono<AuthApiGeneralResponse<AuthApiUserSave>> saveInAuthApi(User user, String appId, String password);
     Mono<AuthApiGeneralResponse<Object>> deleteAuthApiUser(String id);
+    Mono<AuthApiGeneralResponse<Void>> isUserAuthorized(String token);
 }