+++ src/main/java/com/rivalia/rivalia/application/service/UserService.java
@@ -4,11 +4,7 @@ import com.rivalia.rivalia.application.port.in.UserUseCase;
 import com.rivalia.rivalia.application.port.out.UserRepositoryPort;
 import com.rivalia.rivalia.application.port.out.WebClientRepositoryPort;
 import com.rivalia.rivalia.domain.model.User;
-import com.rivalia.rivalia.infraestructure.outbund.webclient.dto.AuthApiGeneralResponse;
-import com.rivalia.rivalia.infraestructure.outbund.webclient.dto.AuthApiUserSave;
 import com.rivalia.rivalia.shared.mapper.GlobalMapper;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 
@@ -25,13 +21,13 @@ public class UserService implements UserUseCase {
     }
 
     @Override
-    public Mono<User> save(User user) {
-        return webClientRepositoryPort.saveInAuthApi(user)
+    public Mono<User> save(User user, String appId, String password) {
+        return webClientRepositoryPort.saveInAuthApi(user, appId, password)
                 .map(authResponse -> {
                     user.setIdAuth(authResponse.getData().getUserId());
                     return user;
                 })
-                .map(userRepositoryPort::save)
+                .flatMap(userRepositoryPort::save)
                 .map(userSaved -> globalMapper.map(userSaved, User.class));
     }
 }