+++ src/main/java/com/rivalia/rivalia/infraestructure/inbound/dto/LoginAppRequest.java
@@ -1,8 +1,11 @@
 package com.rivalia.rivalia.infraestructure.inbound.dto;
 
-import lombok.Builder;
+import lombok.*;
 
 @Builder
+@Getter @Setter
+@AllArgsConstructor
+@NoArgsConstructor
 public class LoginAppRequest {
     private String appName;
     private String password;