+++ src/main/java/com/rivalia/rivalia/infraestructure/outbund/webclient/AuthApiWebClient.java
@@ -5,9 +5,11 @@ import com.rivalia.rivalia.domain.model.User;
 import com.rivalia.rivalia.infraestructure.inbound.dto.LoginAppRequest;
 import com.rivalia.rivalia.infraestructure.outbund.webclient.dto.AuthApiGeneralResponse;
 import com.rivalia.rivalia.infraestructure.outbund.webclient.dto.AuthApiUserSave;
+import com.rivalia.rivalia.infraestructure.outbund.webclient.dto.CreateUserRequestBody;
+import com.rivalia.rivalia.shared.exception.ApiException;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatusCode;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -18,7 +20,7 @@ import java.util.Map;
 @Repository
 public class AuthApiWebClient implements WebClientRepositoryPort {
 
-    private static final String DEFAULT_URI_USERS = "/api/v1/users";
+    private static final String DEFAULT_URI_USERS = "/api/v1/auth/user";
     private static final String APP_ID_VERIFY = "686d406dc46e5ec329ce22ab";
     private static final String APP_NAME = "TESTAPP";
     private static final String APP_PASSWORD = "MTIzNDU2Nzg";
@@ -30,18 +32,32 @@ public class AuthApiWebClient implements WebClientRepositoryPort {
     }
 
     @Override
-    public Mono<AuthApiGeneralResponse<AuthApiUserSave>> saveInAuthApi(User user) {
+    public Mono<AuthApiGeneralResponse<AuthApiUserSave>> saveInAuthApi(User user, String appId, String password) {
         return dynamicWebClient(HttpMethod.POST, DEFAULT_URI_USERS, null, null,
-                null, user, new ParameterizedTypeReference<AuthApiGeneralResponse<AuthApiUserSave>>(){});
+                null, CreateUserRequestBody.builder()
+                        .name(user.getName())
+                        .email(user.getEmail())
+                        .appId(appId)
+                        .password(password)
+                        .lastName(user.getLastName())
+                        .build(), new ParameterizedTypeReference<AuthApiGeneralResponse<AuthApiUserSave>>(){});
     }
 
     @Override
     public Mono<AuthApiGeneralResponse<Object>> deleteAuthApiUser(String id) {
         return loginAppAdmin()
-                .flatMap(loginData -> dynamicWebClient(HttpMethod.DELETE, "/api/v1/app/deleteUser", null
+                .flatMap(loginData -> dynamicWebClient(HttpMethod.DELETE, "/api/v1/auth/app/deleteUser", null
                         ,Map.of("userId", id), Map.of("Authorization", "Bearer " + loginData.getData(), "app-id-verify", APP_ID_VERIFY),
                         null, new ParameterizedTypeReference<>() {}));
     }
+
+    @Override
+    public Mono<AuthApiGeneralResponse<Void>> isUserAuthorized(String token) {
+        return dynamicWebClient(HttpMethod.GET, "/api/v1/auth", null, null,
+                Map.of("Authorization", token, "app-id-verify", APP_ID_VERIFY), null,
+                new ParameterizedTypeReference<>(){});
+    }
+
 
     public Mono<AuthApiGeneralResponse<String>> loginAppAdmin() {
         LoginAppRequest loginAppRequest = LoginAppRequest.builder()
@@ -52,13 +68,14 @@ public class AuthApiWebClient implements WebClientRepositoryPort {
                 loginAppRequest,new ParameterizedTypeReference<AuthApiGeneralResponse<String>>(){});
     }
 
-    public <T, R>Mono<AuthApiGeneralResponse<R>> dynamicWebClient(HttpMethod method,
-                                                            String url,
-                                                            Map<String, ?> pathParams,
-                                                            Map<String, ?> queryParams,
-                                                            Map<String, String> headers,
-                                                            T requestBody,
-                                                            ParameterizedTypeReference<AuthApiGeneralResponse<R>> responseType) {
+    public <T, R> Mono<AuthApiGeneralResponse<R>> dynamicWebClient(HttpMethod method,
+                                                                   String url,
+                                                                   Map<String, ?> pathParams,
+                                                                   Map<String, ?> queryParams,
+                                                                   Map<String, String> headers,
+                                                                   T requestBody,
+                                                                   ParameterizedTypeReference<AuthApiGeneralResponse<R>> responseType) {
+
         WebClient.RequestBodySpec requestSpec = webClient
                 .method(method)
                 .uri(uriBuilder -> {
@@ -75,21 +92,21 @@ public class AuthApiWebClient implements WebClientRepositoryPort {
         }
 
         if (requestBody != null) {
-            return requestSpec
-                    .bodyValue(requestBody)
-                    .retrieve()
-                    .bodyToMono(responseType);
-        } else {
-            return requestSpec
-                    .retrieve()
-                    .onStatus(HttpStatusCode::is4xxClientError, response ->
-                            response.bodyToMono(AuthApiGeneralResponse.class).flatMap(error ->
-                                    Mono.error(new RuntimeException(error.getMessage()))))
-                    .onStatus(HttpStatusCode::is5xxServerError, response ->
-                            Mono.error(new RuntimeException("Server Error")))
-                    .bodyToMono(responseType);
+            requestSpec.bodyValue(requestBody);
         }
+
+        return requestSpec
+                .exchangeToMono(clientResponse -> {
+                    if (clientResponse.statusCode().is2xxSuccessful()) {
+                        return clientResponse.bodyToMono(responseType);
+                    }
+
+                    return clientResponse.bodyToMono(responseType)
+                            .flatMap(errorBody ->
+                                    Mono.error(new ApiException(
+                                            HttpStatus.valueOf(clientResponse.statusCode().value()),
+                                            errorBody
+                                            )));
+                });
     }
-
-
 }